+++ example/start-axum/src/app.rs
@@ -12,6 +12,11 @@ pub fn App() -> impl IntoView {
     // Provides context that manages stylesheets, titles, meta tags, etc.
     provide_meta_context();
 
+    tailwind_fuse::merge::set_tw_options(MergeOptions {
+        prefix: "tw-",
+        separator: ":",
+    });
+
     view! {
         <Stylesheet id="leptos" href="/pkg/start-axum.css"/>
