+++ example/start-axum/src/lib.rs
@@ -7,7 +7,6 @@ pub mod fileserv;
 #[cfg(feature = "hydrate")]
 #[wasm_bindgen::prelude::wasm_bindgen]
 pub fn hydrate() {
-    use crate::app::*;
     console_error_panic_hook::set_once();
-    leptos::mount_to_body(App);
+    leptos::mount_to_body(crate::app::App);
 }
